Update to pattern_patch 0.2.0 and move patch code to PatchHelper
