+++ lib/branch_io_cli/helper.rb
@@ -1,1 +1,2 @@
 require "branch_io_cli/helper/branch_helper"
+require "branch_io_cli/helper/patch_helper"
